fix: Initialize license in queue mode correctly (#6301)
